--- DiscordBot/DiscordBot.Data/Constants.cs
@@ -1,33 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace DiscordBot.Data
-{
-    public static class Constants
-    {
-
-        public enum Languages
-        {
-            Python,
-            JavaScript,
-            Java,
-            Csharp,
-            CPlusPlus,
-            C
-        }
-
-        public enum Levels
-        {
-            Experienced,
-            Intermediate,
-            Beginner
-        }
-
-        public enum UserTypes
-        {
-            Mentor,
-            Mentee
-        }
-    }
-}
