+++ DiscordBot/DiscordBot.Data/Enums.cs
@@ -1,0 +1,25 @@
+﻿namespace DiscordBot.Data
+{
+    public enum Languages
+    {
+        Python,
+        JavaScript,
+        Java,
+        Csharp,
+        CPlusPlus,
+        C
+    }
+
+    public enum Levels
+    {
+        Experienced,
+        Intermediate,
+        Beginner
+    }
+
+    public enum UserTypes
+    {
+        Mentor,
+        Mentee
+    }
+}
