--- DiscordBot/DiscordBot.Data/Models/Mentee.cs
@@ -1,8 +1,0 @@
-﻿namespace DiscordBot.Data.Models
-{
-    public class Mentee
-    {
-
-
-    }
-}
