--- DiscordBot/DiscordBot.Data/Models/Mentor.cs
@@ -1,7 +1,0 @@
-﻿namespace DiscordBot.Data.Models
-{
-    public class Mentor
-    {
-
-    }
-}
