+++ DiscordBot/DiscordBot.Data/Models/User.cs
@@ -10,7 +10,8 @@ namespace DiscordBot.Data.Models
         }
 
         public ulong UserId { get; set; }
-        public virtual IList<UserLanguages> UserLanguages { get; set; }
+        public UserTypes UserType { get; set; }
+        public virtual IList<UserLanguage> UserLanguages { get; set; }
 
         public override string ToString()
         {
