+++ DiscordBot/DiscordBot.Data/Repositories/IMentorRepository.cs
@@ -9,11 +9,11 @@ namespace DiscordBot.Data.Repositories
     {
         int GetCount();
         //List<Programmer> GetMentors(int begin, int end);
-        Task<IEnumerable<Mentor>> GetMentorsAsync();
-        Task<Mentor> GetMentorAsync(ulong id);
-        Task InsertMentorAsync(Mentor mentor);
-        Task DeleteMentorAsync(ulong id);
-        Task UpdateMentorAsync(Mentor mentor);
+        //Task<IEnumerable<Mentor>> GetMentorsAsync();
+        //Task<Mentor> GetMentorAsync(ulong id);
+        //Task InsertMentorAsync(Mentor mentor);
+        //Task DeleteMentorAsync(ulong id);
+        //Task UpdateMentorAsync(Mentor mentor);
         Task SaveAsync();
     }
 }
